requirements.txt
--prefer-binary
fastapi==0.103.2
sqlmodel==0.0.10
alembic==1.12.0
python-dotenv==1.0.0
email-validator<2.0.0
httpx==0.25.0
orjson==3.9.9
inflection==0.5.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
psycopg2-binary==2.9.9
asyncpg==0.28.0
redis<5.0.0
minio==7.1.17
pillow==10.1.0
authlib<2.0.0==None
itsdangerous==2.1.2
fastapi-mail==1.3.1
uuid6==2023.5.2
fastapi-pagination==0.12.10
hypercorn==0.14.4
bcrypt==4.0.1
fastapi-cache2[redis]==0.2.1
cryptography==41.0.4
siwe==2.2.0
petname==2.6.0
